haciendo el primer proyecto desde 0

+++ My project/Assets/Cartas/IncreaseCard.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class IncreaseCard : BaseCard
 {
-    public double Incremento;
-    public IncreaseCard(string Name, int InitialPower, Faction Faction, TipoDeCarta TipoDeCarta, List<Zonas> destinations, double Incremento = 0, EffectDelegate Effect = null) : base(Name, InitialPower, Faction, TipoDeCarta, destinations, Effect)
+    public int Incremento;
+    public IncreaseCard(string Name, int InitialPower, Faction Faction, TipoDeCarta TipoDeCarta, List<Zonas> destinations, int Incremento = 0, EffectDelegate Effect = null) : base(Name, InitialPower, Faction, TipoDeCarta, destinations, Effect)
     {
         this.Incremento = Incremento == 0? 1: Incremento;
     }
